+++ Tests/MapboxDirectionsTests/AttributeOptionsTests.swift
@@ -7,10 +7,10 @@ class AttributeOptionsTests: XCTestCase {
         var options2merge = AttributeOptions(descriptions: ["speed"])!
         var optionsWithCustom = AttributeOptions()
 
-        optionsWithCustom.update(customOption: (1<<7, "Custom7"), comparisonPolicy: .allowEqual)
+        optionsWithCustom.update(customOption: (1<<7, "Custom7"), comparisonPolicy: .equal)
         options.update(with: .distance)
         options.update(with: optionsWithCustom)
-        options2merge.update(customOption: (1<<8, "Custom_8"), comparisonPolicy: .allowEqual)
+        options2merge.update(customOption: (1<<8, "Custom_8"), comparisonPolicy: .equal)
 
         options.update(with: options2merge)
         
@@ -20,7 +20,7 @@ class AttributeOptionsTests: XCTestCase {
         XCTAssertEqual(options.description.split(separator: ",").count,
                        4)
         XCTAssertEqual(optionsWithCustom,
-                       options.update(customOption: (1<<7, "Custom7"), comparisonPolicy: .allowEqual))
+                       options.update(customOption: (1<<7, "Custom7"), comparisonPolicy: .equal))
         
         // insert existing default
         XCTAssertFalse(options.insert(.distance).inserted)
@@ -28,8 +28,8 @@ class AttributeOptionsTests: XCTestCase {
         XCTAssertFalse(options.insert(optionsWithCustom).inserted)
         // insert conflicting custom
         var optionsWithConflict = AttributeOptions()
-        optionsWithConflict.update(customOption: (optionsWithCustom.rawValue, "Another custom name"), comparisonPolicy: .allowEqual)
-        XCTAssertFalse(options.insert(optionsWithConflict, comparisonPolicy: .allowUnequal).inserted)
+        optionsWithConflict.update(customOption: (optionsWithCustom.rawValue, "Another custom name"), comparisonPolicy: .equal)
+        XCTAssertFalse(options.insert(optionsWithConflict, comparisonPolicy: .rawValueEqual).inserted)
         // insert custom with default raw
         optionsWithConflict.rawValue = AttributeOptions.distance.rawValue
         XCTAssertFalse(options.insert(optionsWithConflict).inserted)
@@ -38,29 +38,29 @@ class AttributeOptionsTests: XCTestCase {
     func testContains() {
         var options = AttributeOptions()
         options.update(with: .expectedTravelTime)
-        options.update(customOption: (1<<9, "Custom"), comparisonPolicy: .allowEqual)
+        options.update(customOption: (1<<9, "Custom"), comparisonPolicy: .equal)
         
         XCTAssertTrue(options.contains(.init(rawValue: AttributeOptions.expectedTravelTime.rawValue)))
         XCTAssertFalse(options.contains(.congestionLevel))
         
         var wrongCustomOption = AttributeOptions()
-        wrongCustomOption.update(customOption: (1<<9, "Wrong name"), comparisonPolicy: .allowEqual)
+        wrongCustomOption.update(customOption: (1<<9, "Wrong name"), comparisonPolicy: .equal)
         XCTAssertFalse(options.contains(wrongCustomOption))
         
         var correctCustomOption = AttributeOptions()
-        correctCustomOption.update(customOption: (1<<9, "Custom"), comparisonPolicy: .allowEqual)
+        correctCustomOption.update(customOption: (1<<9, "Custom"), comparisonPolicy: .equal)
         XCTAssertTrue(options.contains(correctCustomOption))
         
-        XCTAssertTrue(options.contains(.init(rawValue: 1<<9), comparisonPolicy: .allowEqualOrNull))
+        XCTAssertTrue(options.contains(.init(rawValue: 1<<9), comparisonPolicy: .equalOrNull))
     }
     
     func testRemove() {
         var preservedOption = AttributeOptions()
-        preservedOption.update(customOption: (1<<12, "Should be preserved"), comparisonPolicy: .allowEqual)
+        preservedOption.update(customOption: (1<<12, "Should be preserved"), comparisonPolicy: .equal)
         var options = AttributeOptions()
         options.update(with: .congestionLevel)
         options.update(with: .distance)
-        options.update(customOption: (1<<10, "Custom"), comparisonPolicy: .allowEqual)
+        options.update(customOption: (1<<10, "Custom"), comparisonPolicy: .equal)
         options.update(with: preservedOption)
         
         // Removing default item
@@ -77,7 +77,7 @@ class AttributeOptionsTests: XCTestCase {
         
         // Removing custom option with incorrect name
         var wrongCustomOption = AttributeOptions()
-        wrongCustomOption.update(customOption: (1<<10, "Wrong name"), comparisonPolicy: .allowEqual)
+        wrongCustomOption.update(customOption: (1<<10, "Wrong name"), comparisonPolicy: .equal)
         
         XCTAssertNil(options.remove(wrongCustomOption))
         XCTAssertTrue(options.contains(.congestionLevel))
@@ -85,7 +85,7 @@ class AttributeOptionsTests: XCTestCase {
         
         // Removing existing custom option
         var correctCustomOption = AttributeOptions()
-        correctCustomOption.update(customOption: (1<<10, "Custom"), comparisonPolicy: .allowEqual)
+        correctCustomOption.update(customOption: (1<<10, "Custom"), comparisonPolicy: .equal)
         
         XCTAssertEqual(options.remove(correctCustomOption), correctCustomOption)
         XCTAssertTrue(options.contains(.congestionLevel))
@@ -93,12 +93,12 @@ class AttributeOptionsTests: XCTestCase {
         
         // Removing custom option with default raw value
         var customOptionWithDefaultRaw = AttributeOptions()
-        customOptionWithDefaultRaw.update(customOption: (AttributeOptions.distance.rawValue, "Not a distance"), comparisonPolicy: .allowEqual)
+        customOptionWithDefaultRaw.update(customOption: (AttributeOptions.distance.rawValue, "Not a distance"), comparisonPolicy: .equal)
         XCTAssertNil(options.remove(customOptionWithDefaultRaw))
         
         // Removing custom option by raw value only
         options.update(with: correctCustomOption)
-        XCTAssertEqual(options.remove(.init(rawValue: 1<<10), comparisonPolicy: .allowEqualOrNull), correctCustomOption)
+        XCTAssertEqual(options.remove(.init(rawValue: 1<<10), comparisonPolicy: .equalOrNull), correctCustomOption)
     }
     
     func testCustomAttributes() {
@@ -107,8 +107,8 @@ class AttributeOptionsTests: XCTestCase {
         var attributes = AttributeOptions()
         attributes.insert(.congestionLevel)
         attributes.insert(.speed)
-        attributes.update(customOption: customOption1, comparisonPolicy: .allowEqual)
-        attributes.update(customOption: customOption2, comparisonPolicy: .allowEqual)
+        attributes.update(customOption: customOption1, comparisonPolicy: .equal)
+        attributes.update(customOption: customOption2, comparisonPolicy: .equal)
         
         let descriptions = attributes.description.split(separator: ",")
         XCTAssertTrue(descriptions.contains { $0 == AttributeOptions.congestionLevel.description })